Implement a deep sleep mode
It disables pretty much everything except the async pin change interrupts,
and should bring the AVR down to the 1µA range, in theory.

Note: BOD sleep wasnt planned to be used with these fuse settings,
and officially it might not come out of the sleep sanely - but
i dont really care if it ends up resetting or something, and well
we are disabling the brownout detector anyways... in my tests
it seemed to work fine, but i dont even know if it's implemented
on the rev of my ATtiny85V lol - i should test, and also i should
measure the current ... but my accurate multimeter is AWOL, sigh.